Add the makings of a home page
While it doesn't do anything yet, this is the beginning of what the home
page will be. It includes two placeholder anchors that will eventually
link to routes that will allow people to apply for financial aid and
scholarships.

Since it merely served as a template, the `hello` endpoint is being
removed.