Use Tailwind as CSS framework
This is mostly experimental at this point. If we're happy with Tailwind
as a framework and we end up needing customization, we'll need to
include it locally and add it to our build process (we isn't even a
thing yet), but for now the CDN version will suffice.

As for the styling being applied, the two anchors will appear stacked on
small screens and side by side on medium (and up) screens.