Add a model for tracking email addresses
The `User` class will be used to store email addresses associated with
applications. It can eventually be used to store email addresses that
belong to people who can approved the applications.

This also makes how [Alembic](https://alembic.sqlalchemy.org) accesses
the metadata consistent with the comment about it.

References #4