Use Alembic for migrations
[Alembic](https://alembic.sqlalchemy.org) is a library for managing
migrations for [SQLAlchemy](https://www.sqlalchemy.org) tables. There
are no tables to migrate yet, but this will add the initial setup. A
couple of modifications are being made to the default template: the
application's `Base` class's metadata is being provided to Alembic and
the current working directory is being added to `sys.path` so that
Alembic can import from `awards`.

In addition to adding Alembic, a new tox environment is being added to
test that migrations are forwards and backwards compatible. This is
being added to GitHub Actions as a new job.