Check that migrations have been generated
This will add a [pre-commit](https://pre-commit.com) hook to run
[alembic-autogen-check](https://pypi.org/p/alembic-autogen-check) to
ensure that the Alembic migrations are up-to-date. It will run whenever
a Python file is being committed. In the future, we may want to consider
limiting the scope to files that contain model definitions.

Unfortunately there's no way to get pre-commit to install from a
`requirements.txt` file (see
https://github.com/pre-commit/pre-commit/issues/730#issuecomment-374463792),
so [tox](https://pypi.org/p/tox) is being used to run
`alembic-autogen-check`.