Only check migrations are up-to-date in CI
It turns out that checking if the
[Alembic](https://alembic.sqlalchemy.org] are up-to-date can be a
relatively expensive operation, at least compared to what you typically
want to see when using [pre-commit](https://pre-commit.com) hooks. Not
only do you need to install all of the dependencies, but you also need
to run all of your migrations first. On top of that, because our
dependencies can't reasonably be enumerated in our
[tox](https://pypi.org/p/tox) configuration, the environment needs to be
recreated on each run.

The trade off here is that no one will receive feedback about missing
migrations until CI. As the sole contributor to the project at this
time, though, that's a trade off I'm okay with. If future contributors
ever see a non-trivial number of CI failures because of this, we can
revisit this approach.

Since the two tox environments will be invoked in the same CI step, they
are being renamed so that the intention of each is clearer.

Closes #11