Use Postgres as the database
While the plan had always been to use
[Postgres](https://www.postgresql.org) as the database, the goal was to
introduce it as late as possible to keep the developer experience as
simple as possible for as long as possible. Unfortunately the migration
tests in #13 are failing because the database doesn't persist across
commands in the [tox](https://tox.readthedocs.io) environments.

Rather than spending time making changes to the tox environments just to
try to stick to in-memory [SQLite](https://www.sqlite.org) for a little
longer, this will move the project over to Postgres right now. The
`check-migrations` environment is being merged with the
`test-migrations` environment to simplify managing them. They're now
known as `migrations`.

Each tox environment that requires a database will look for a specific
environment variable that it will then pass along to the application as
`DATABASE_URI`.