Use a better name for the unit tests environment
This project doesn't need to support multiple versions of Python.
There's no need to name a tox environment after one just because that's
a thing you can do. By renaming the environment from `py38` to `unit`,
not only do we better convey what the environment is testing, but it's
also one less thing that has to be changed when we upgrade to a newer
version of Python.