Switch to Django
While I greatly prefer [SQLAlchemy][sqlalchemy] and its
[ORM][sqlalchemy-orm] to [Django's ORM][django-orm], this project most
likely won't involve any complex database queries. A good portion of the
time spent on this project has been getting things to work in such a way
that both [Flask][flask] and [Mypy][mypy] are happy. I don't really have
time for that. Conferences have very real deadlines and [Django][django]
is all about deadlines.

> The web framework for perfectionists with deadlines.

As such, I'm switching the project over to use Django. This isn't an
exact 1:1 translation of the functionality that is supported by the
existing implementation thus far, but it was done in a lot less time and
should allow for completing the project much quicker.

In addition to switching to Django, Mypy is being dropped in favor of
[Pyre][pyre]. The advantage of Pyre is that it ships with stubs for
Django. To keep Mypy, we'd need to a plugin, the most popular of which
is [django-stubs][django-stubs]. Unfortunately django-stubs executes
some of the code, including the settings. This would require providing
environment settings for anything configurable. Furthermore, the type
checks are no longer going to run under [pre-commit][pre-commit]. As
with mypy, you get much better results when the dependencies are
installed. They will instead run as a separate CI job under [tox][tox].

[sqlalchemy]: https://www.sqlalchemy.org
[sqlalchemy-orm]: https://docs.sqlalchemy.org/en/13/orm/
[django-orm]: https://docs.djangoproject.com/en/3.0/topics/db/
[flask]: https://flask.palletsprojects.com/en/1.1.x/
[mypy]: https://mypy.readthedocs.io
[django]: https://www.djangoproject.com
[pyre]: https://pyre-check.org
[django-stubs]: https://pypi.org/p/django-stubs
[pre-commit]: https://pre-commit.com
[tox]: http://tox.readthedocs.io