Add tests for logging in
For the most part I trust `LoginForm` and [django-sesame][django-sesame]
to work, but it's good to make sure that our custom login behavior
works, namely that a new `User` will be created whenever a new email
address is submitted. This will also test that forms without an email
addressed don't validate.

[django-sesame]: https://django-sesame.readthedocs.io