Accept applications for tickets
This is the beginning of the `applications` app. While it will
eventually do more, this will add an endpoint (`scholarship`) that
allows a user to apply for a ticket (a.k.a. a scholarshipship). The
application is validated by the `ScholarshipApplicationForm` and saved
to the `Application` model.

Before submitting an application, a user will need to be logged in.