Redact email addresses
The default string representation of a `User` is their email address.
Even though access to the data will be restricted, it's a bad idea to
display someone's email address on a web page. It could also lead to
bias when reviewing their application as you may see their name in their
email address.

We may need to consider redacting the domains as well but for now this
is a good start.