Assign errors to specific financial aid form fields
When raising a [`ValidationError`][validationerror], it's common to
provide it with a single string containing the error message. It turns
out it can also take a list of strings. Just as with a single string,
the list will be handled as form-level errors.

`ValidationError` accepts a third kind of message: a dictionary. When
providing a dictionary, the keys represent the names of the fields with
which the errors should be associated.

`FinancialAidApplicationForm.clean` can utilize this behavior to better
associate the errors with their fields. This will also simplify the
tests.

I'm removing the checks for the specific error text as this tends to be
brittle. I did catch a missing "." at the end of a message when I wrote
one, but our checks are pretty straightforward, so hopefully the simpler
checks are good enough.

[validationerror]: https://github.com/django/django/blob/d6aff369ad33457ae2355b5b210faf1c4890ff35/django/core/exceptions.py#L101-L108