Improve documentation for tox environments
Several of the [tox][tox] environments exist to make it easier to work
with the project (as opposed to just testing it). While a couple of them
were covered in the documentation, most of them were not. This adds new
sections to the README to make these environments more discoverable.

[tox]: https://tox.readthedocs.io