Make it easier to use a debugger with runserver
This introduces two enhancements to the `runserver` [tox][tox]
environment to make it easier to debug the application. First, it passes
the [`PYTHONBREAKPOINT` environment variable][breakpoint] into the
environment, allowing users to take advantage of the `breakpoint`
built-in. Second, it installs [bpdb][bpdb], [ipdb][ipdb], and
[pudb][pudb] into the environment to allow people to use the debugger
they're most comfortable using (similar to how the interface can be
controlled when using the `shell` environment).

[tox]: https://tox.readthedocs.io
[breakpoint]: https://www.python.org/dev/peps/pep-0553/#environment-variable
[pudb]: https://pypi.org/p/pudb
[bpdb]: https://bpython-interpreter.org
[ipdb]: https://pypi.org/p/ipdb