Upgrade all of the requirements
[Django 3.0.7][django_3_0_7] is a security release. Rather than
upgrading it and leaving everything else alone, I'm taking the
opportunity to upgrade all of the dependencies listed in the project's
requirements files.

I did this by added `--upgrade` as an argument to each usage of the
`pip-compile` [pre-commit][pre-commit] hook. I then ran them using

    pre-commit run pip-compile --all-files

After that I discarded the changes to `.pre-commit-config.yaml`.

[django_3_0_7]: https://www.djangoproject.com/weblog/2020/jun/03/security-releases/
[pre-commit]: https://pre-commit.com