Make applications viewable by the applicant
Currently applications can see the list of applications they've
submitted, but they can't actually view their details. This adds a new
`view` view to the `applications` app that will allow them to view their
own, but not other users', applications.

[factory_boy][factory_boy] is being used to create model instances for
the tests. It may be worth investigating a more reusable approach to
defining the factories.

Closes #49

[factory_boy]: https://factoryboy.readthedocs.io