Fix links in README
These bits of text were intended to be links (which is why they're
included in the URL block at the bottom of the README). This correctly
identifies them as such.

While #67 also tried to do this, the intention of the text being
preformatted to indicate values that could be passed to the command was
lost. This expresses that better.