Add a note about Hacktoberfest
To reduce the burden on maintainers and to help combat spam,
[Hacktoberfest] has gone opt-in for 2020. As someone who has happily
participated in this event in the past, I'm accepting pull requests to
this project. This adds a note to let people know that spam PRs will be
rejected and I won't hesitate to pull the plug if it becomes a problem.

[hacktoberfest]: https://hacktoberfest.digitalocean.com