Update the version of dependencies
The only dependency not being updated here, at least not fully, is
[Django]. That's going to be done separately as part of #80. For now the
version is being fixed to the latest version of 3.0.

The Django version also has to be set for the development requirements.
Without doing so, the latest version would be pulled in by
[django-debug-toolbar]. This would lead to conflicting versions in CI as
some of the [tox] environments install both the development and regular
dependencies.

[django]: https://www.djangoproject.com
[django-debug-toolbar]: https://pypi.org/p/django-debug-toolbar
[tox]: https://tox.readthedocs.io